requirements.txt
# airflow packages
alembic==1.7.5
apache-airflow-providers-google==6.4.0
apache-airflow-providers-slack==4.2.0
apache-airflow[postgres,jdbc,mysql,google,google_auth,cncf.kubernetes,ssh,sftp,statsd]==2.3.1rc1
Authlib==0.15.5
elementpath==2.2.3

Flask-SQLAlchemy==2.5.1
gcsfs==2022.1.0
google-api-core==1.31.5
google-api-python-client==1.12.10
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery-datatransfer==3.6.0
google-cloud-bigquery-storage==2.11.0
google-cloud-bigquery==2.33.0
google-cloud-storage==1.44.0
iso8601==1.0.2
jsondiff==1.3.1
jsonpath-ng==1.5.3
lxml==4.8.0
marshmallow-sqlalchemy==0.26.1
oss2==2.15.0
pandas-gbq==0.14.1
pdpyras==4.4.0
pyarrow==6.0.1
pyasn1==0.4.8
pymongo==3.12.3
pyOpenSSL==20.0.1
pytz==2021.3
requests==2.27.1
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
SQLAlchemy==1.4.9
xmltodict==0.12.0
pyzipper==0.3.5

# custom packages
fastparquet
great_expectations
ndg-httpsclient
openpyxl
PGPy
pybigquery==0.5
pybuilder
pyminizip==0.2.6
setuptools
wheel
xlrd

# Unit test requirement
pytest-cov==2.12.1
pytest==6.2.5
